kopf-example-restarter-operator/app/requirements.txt

aiohttp==3.7.4.post0
aiojobs==0.3.0
astroid==2.3.3
async-timeout==3.0.1
attrs==21.2.0
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
coverage==5.5
dictdiffer==0.8.1
google-auth==1.30.1
idna==2.10
importlib-metadata==4.5.0
iso8601==0.1.14
isort==4.3.21
kopf==1.32.1
kubernetes==12.0.0a1
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.8.0
multidict==5.1.0
oauthlib==3.1.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.4.4
pyparsing==2.4.7
pytest==5.4.3
pytest-cov==2.8.1
pytest-env==0.6.2
pytest-mock==2.0.0
python-dateutil==2.8.1
python-json-logger==2.0.1
PyYAML==5.4
requests==2.25.1
requests-oauthlib==1.3.0
rope==0.16.0
rsa==4.7.2
six==1.16.0
typed-ast==1.4.3
typing-extensions==3.10.0.0
urllib3==1.26.5
wcwidth==0.2.5
websocket-client==1.0.1
wrapt==1.11.2
yarl==1.6.3
zipp==3.4.1
environs==9.3.2 